refactored CSV import components into a single custom component
