+++ apps/frontend/src/components/ViewFormsList/ViewFormsList.tsx
@@ -14,15 +14,9 @@ import {
   Thead,
   Tr,
   Input,
-  Text,
-  Button,
 } from '@chakra-ui/react';
 import { useNavigate } from 'react-router-dom';
-import {
-  getAllForms,
-  markFormAsRead,
-  putMultipleCSVForms,
-} from '../../utils/sendRequest';
+import { getAllForms, markFormAsRead } from '../../utils/sendRequest';
 import { FormData } from '../../types/formData';
 import { SortOptions, SortOrder } from '../../enums/SortOrder';
 import useFormsListFiltering from '../../hooks/useFormsListFiltering';
@@ -31,10 +25,9 @@ import {
   lastUpdatedCompareFunction,
   nameCompareFunction,
 } from '../../utils/sortFunctions';
-import CSVImportButton from './CSVImportButton';
+import CSVImport from './CSVImport';
 
 export default function ViewFormsList() {
-  const [uploadedCSV, setUploadedCSV] = useState<File | null>(null);
   const [forms, setForms] = useState<FormData[]>([]);
   const [allForms, setAllForms] = useState<FormData[]>([]); // this is used to get all forms again after removing a filter/search term
   const [sortBy, setSortBy] = useState<SortOptions>(SortOptions.NAME);
@@ -89,65 +82,12 @@ export default function ViewFormsList() {
     getForms();
   }, []);
 
-  // called when the user uploads a new file via the "choose file" button
-  // stores the uploaded file if it is a CSV file, or alerts the user
-  // if they uploaded the wrong file type
-  function handleFileUpload() {
-    const csv_input = document.getElementById('csv_upload') as HTMLInputElement;
-    const files = csv_input.files;
-    setUploadedCSV(null);
-    if (files?.length === 1) {
-      const uploadedFile = files[0];
-      if (uploadedFile.type === 'text/csv') {
-        setUploadedCSV(uploadedFile);
-      } else {
-        alert('Please upload a CSV file.');
-        csv_input.value = '';
-      }
-    }
-  }
-
-  // calls the endpoint to import CSV form data if a CSV file has been
-  // uploaded, or alerts the user otherwise
-  async function handleImportButtonClick() {
-    if (uploadedCSV === null) {
-      alert('Please upload a CSV file first.');
-    } else {
-      putMultipleCSVForms(await uploadedCSV?.text());
-    }
-  }
-
   return (
     <Box p={1}>
       <Center mb={1}>
         <Heading size="xl">Submitted Forms</Heading>
       </Center>
-      <Box
-        display="flex"
-        flexDirection="column"
-        alignItems="start"
-        rowGap="5px"
-        mb="10px"
-        marginLeft="auto"
-        marginRight="auto"
-        width="98%"
-      >
-        <Text>Upload a CSV file to import completed forms:</Text>
-        <input
-          id="csv_upload"
-          name="csv_upload"
-          type="file"
-          accept=".csv"
-          onChange={handleFileUpload}
-        ></input>
-        <Button
-          id="csv_import"
-          name="csv_import"
-          onClick={handleImportButtonClick}
-        >
-          Import
-        </Button>
-      </Box>
+      <CSVImport />
       <Box display="flex" flexDirection="row" justifyContent="space-between">
         <Select
           width="25%"
